+++ app/src/main/java/com/concordium/wallet/ui/cis2/TokenDetailsActivity.kt
@@ -83,6 +83,7 @@ class TokenDetailsActivity : BaseActivity() {
             setContractIndexAndSubIndex(token)
             setTokenId(token.token)
             token.tokenMetadata?.let { tokenMetadata ->
+                setName(tokenMetadata)
                 setImage(tokenMetadata)
                 setOwnership(token, tokenMetadata)
                 setDescription(tokenMetadata)
@@ -154,6 +155,11 @@ class TokenDetailsActivity : BaseActivity() {
                 binding.includeAbout.contractIndex.text = tokenIndex
             }
         }
+    }
+
+    private fun setName(tokenMetadata: TokenMetadata) {
+        binding.includeAbout.nameAndIconHolder.visibility = View.VISIBLE
+        binding.includeAbout.name.text = tokenMetadata.name
     }
 
     private fun setImage(tokenMetadata: TokenMetadata) {